Refactors home view to improve user experience
Refactors the main view to provide a better user experience by displaying different content based on whether the user has planned their week or not.

Improves the layout and information shown on the home screen.
Adds a slider to initiate the workout.

Also makes minor adjustments to navigation and workout actions.